moved `_u` and `_v` to shared.py
