+++ jovian/tests/resources/__init__.py
@@ -1,1 +1,1 @@
-from jovian.tests.resources.common import fake_creds
+from jovian.tests.resources.shared import fake_creds
